Welcome Cleo and Jasper to the cast!
Cleo and Jasper are the final two cast members. Cleo was hard to write a basis for, because she is not personified BY the rage beneath the surface, but we'll certainly see it bubble up when interacting with her, and see her keeping it under wraps, perhaps to her detriment (taking some inspo from Life is Strange True Colours' approach to anger as something that needs and demands to be felt). Jasper (can you tell there's a chef in my life?) might need a little more nuance, but as a foundation I like the idea of a helping hand as a procrastinator, but also the adjustment to living alone and away from home is a hard one and it becomes very overwhelming very quickly!

Onto story beats, so that I can put this to the Behaviour Group to go over tomorrow.