Overarching Story Beats Added
I've set up the skeleton of the story, but most of it will now boil down to the individual stories between the PC and the Other (I forgot that was what I was calling them in the prelim doc).

What I am bumping into is that dating sims tend to let you focus on one bachelor/bachelorette, where in this game I need the PC to be able to interact with all the cast to give them ample opportunity for success in getting folks to sign onto the petition (though, I need to think about what the threshold is for that... and what it represents...).

With the sinus infection last weekend into this week, I'm a day or so behind what I wanted to send to the team, so I might send off some late-night workings that if folks get the chance to look over, but don't expect them to look over. Or even just show them in the meeting tomorrow - specifically the visuals, I wanted the basic narrative and characters to the team.