Create and uploading LYNDA_SCENE2.md
This is the test scene I've written for my demo this week at the Behaviour Meeting. I've yet to implement it into the engine, but I figured it was worth having a separate file to update the story edits that will then be implemented in engine. It'll help me keep track of any change I do or miss when adding it into Ren'Py as well. 

A big part of writing this was ensuring the theme of impostor syndrome was present without bashing the player over the head with it. We'll see how successful that is with playtesting, and whether folks remember the character traits or if they forget whether they can trace it in the scene. Anyway, to ENGINE!