Create and updated MEET_THE_TEAM.md
Here's the next script I'm writing that I'll port into RenPy. This is the "meet the cast" day one script. It'll be a longer one to cover meeting all four characters and establishing some player skills along the way. I am still toying with a more Disco Elysium style character skill set (within reason) to shape them beyond solely in situ character choices (and also that the skills might inform certain dialogue options). I need to make a tab with all the Skills I'm putting in place. 

Also - when labelling these skills I'm thinking a lot about what they represent. Like, in this document we see ShareSkill which I initially had as SocialSkill, but found that to be too broad. So instead this is a marker of how much the player might overshare in certain moments, about themselves or other characters. Now I'm wondering if the PC overshares it can maybe allow certain characters to open up to them more and make others more hesitant. So that there's a benefit to oversharing. Hm.