Update MEET_THE_TEAM.md
Not quite done yet, I've added Antony and Sasha (with a hint of Lynda) to the first day.

I also realised that I was making the choices action-based rather than dialogue based. This is a really interesting tension to explore in how choice-systems frame their choices. Especially in the space between "show don't tell" in writing practices and "expressing feeling" in emotional game design. In order to let the player express what they want in the moment I wonder if I need to label the choices (such as the change here between Share and Overshare being the options where I've now added dialogue and labelled the dialogue as such).

I was hoping to have the whole script for this day written by now, but here we are. Will try to do a little more this evening so it is complete before folks might read it tomorrow.