Create LYNDA_INTERACTION3
The second of two newly written scenes that I will be adding to engine this week. Annoyingly, the markdown converter doesn't include my comments as I've been writing, so I want to draw a couple to highlight some of my process:

"I'm enjoying some of the choices just being a nod to mental states and decisions we make in conversation, building off of folks' feedback that it not being an "exhaust all options" flow is a nice change and holds up the intention of the game (to get to know and understand these tenants) well."
- This is in relation to making the dialogue feel like it has a more natural flow, and that deciding to dip from conversation (in this case it was choosing to not pry more into Lynda's feelings) feels justified rather than a big blinking "END DIALOGUE" vibe to the choice.

"also thoughts here about whether i soft lock players out of interactions if they have a low enough rating. for this as a vertical slice and all. maybe folks can still save themselves if they don't do interaction 2 (helping Lynda with Tech) and I just have to math out the relationship numbers for the outcome."
- I want there to be a fail state. If a player "doesn't care" or get to know the NPC then perhaps the relationship degrades into acquaintanceship where they see Lynda around less, or they react differently to their presence (like a VERY downscaled version of BG3's reputation system).

I also (in writing this scene) am thinking more about the art side of it. How the emotions of Lynda will be displayed so I won't always have to write it out, but with limited asset creation I still want to include some descriptions in places such as "she wipes at her smudged eyeliner" rather than "lynda smiles". 

Expect some reflections throughout the week, and an incoming file dump from Seb helping me figure out that I *can* in fact not write this as one gigantic file!